Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > > Patch Set 1:
> > > > > 
> > > > > > Patch Set 1:
> > > > > > 
> > > > > > Ooops, hit enter too fast.
> > > > > > 
> > > > > > What's the plan for keeping that list up-to-date? If we're doing something source-based, we'll definitely need one, since we can't reasonably manually update that. It would also be preferable if generating the list were automated so we don't have a lot of diff noise every time we add a function.
> > > > > 
> > > > > I was actually thinking that we'd require manual updating, but provide an automated test to compile two versions of the build artifacts with two different prefixes and then compare them to make sure that there's no overlap between the symbols. That way, you'd always catch people who forgot, and once you know to update the list, doing so takes 30 seconds.
> > > > 
> > > > That still sounds like a huge amount of friction on development. Post-facto testing allows moving from "automated" to "mostly automated", but I think the "mostly" is still important.
> > 
> > Where do you imagine the friction coming in? Presumably, given a command to run the test, a developer could do this along with whatever other testing they're already doing, so they wouldn't have to wait for a cycle through CQ to discover the issue.

No, the list of symbols is platform-dependent, so it wouldn't happen along with whatever testing that's already happening. It would involve CQ cycles in the general case. This also doesn't currently work if one is developing on Windows.

And when the testing says something went wrong, needing to go back and manually edit a bunch of free-form macros is a huge pain. There is also no plan in this CL for clearing out old macros when functions are renamed.

> > Alternatively, how common are CLs that add new functions vs update existing ones? I looked through the diffs of the most recent 10 CLs, and none of them added new functions. So a check in CQ might not add much friction in the common case.

Banking on not adding functions often is definitely not okay. That disincentives cleaning up code. There used to be function codes in the error system which had exactly this problem. We got rid of them. This CL is even worse because the list is not scripted.

> Actually, better yet: If there won't be many consumers of this feature (probably just me for the forseeable future), maybe you could put it in but not test it in CQ and not guarantee that it's kept up to date? The tests would live in-tree, but wouldn't be run in CQ. I could take the responsibility of running the tests when I vendor a new version, and if the tests fail, I would be forced to submit a CL to fix the problem before I could vendor. It'd slow me down a bit, but not badly, and it'd have 0 impact on your development workflow.

If we carry random things in our tree that are perennially out-of-date, as soon as we get a second consumer, there will be pressure to test it and keep it up-to-date continuously, so this doesn't add anything. If we want to defer to when you update your vendored version, then just keep the symbol list in your tree. It's the same amount of work for you.

Long-term, yeah, this should be more broadly automated and supported directly. There's a number of things one could explore there. Maybe we have a source file parser plus an overrides list for asm. But you need a renaming thing now, no?

So we should figure out the minimal-overhead mechanism necessary to get you there and we iterate on it over time.

> > > Here's a concrete proposal.
> > > 
> > > 1. Add a tool that takes a list of symbols and generates those header files. That's probably a good idea anyway, given that __APPLE__ + asm requires something funny. Later on, Windows will require generating things in nasm format.
> > > 
> > > 2. Rather than check in the files, make CMake take both -DBORINGSSL_PREFIX=whatever -DBORINGSSL_PREFIX_SYMBOLS=path/to/file/list. Then CMake generates those headers and whatnot.
> > > 
> > > 3. In your build:
> > > 
> > > 3a. Build BoringSSL without -DBORINGSSL_PREFIX.
> > > 3b. Extract the symbols out and make a symbol list. Skip any C++ symbols (start with "_Z")
> > > 3c. Build BoringSSL again, passing that symbol list in.
> > > 3d. Compare the two didn't overlap in symbols.
> > > 3e. Optional: do the same with a different -DBORINGSSL_PREFIX.
> > > 
> > > If I understand your scenario correctly, that should work for you, right? 
> > 
> > I take it you're suggesting doing this every time you build? Seems like a lot of complexity and risk for introducing errors... Not to mention, it'd slow an already-slow build down by 2x.

Eh?

$ time ninja -C build-release crypto ssl
ninja: Entering directory `build-release'
[297/297] Linking CXX static library ssl/libssl.a

real	0m12.862s
user	1m24.994s
sys	0m6.829s

Perhaps we have different standards for slow. I come from Chromium. :-)

You could also do (3a) and (3b) when you update your vendored version, but then you *really* do need to do (3d) as part of your build because of configuration-specific symbols. By far the most robust answer will be to do (3a) and (3b) as part of the build itself.

But if you don't like building it twice, you can manage the symbol list however you like, and then replace (3d) with checking that all the symbols contain the prefix as a substring. (It needs to be a substring check because of C++ symbols.)

All this is what I meant when I said the macro option had the most unknowns. If you go with the macro strategy, you will need to accept extra overhead while we figure out those unknowns.

> > > It's a bit more annoying for, say, WebRTC, but we have to solve the HMAC symbol for anyone else anyway. If we end up needing to do that, I'm thinking...
> > > 
> > > - Symbols like HMAC split into static inline ... HMAC(...) { return CRYPTO_hmac(...); }
> > > - Macros only perturb CRYPTO_hmac.
> > > - We come up with some source-only story for pre-generating the symbol lists. Like maybe we parse out the headers and have an extras file for any symbols this misses. Assembly, platform-specific ifdefs, etc.
> > > 
> > > > > > The C++ bits need an inline namespace, of course, but we can do that separately. That one's really easy.
> > 
> > Do we need to handle C++? Isn't it the case that the macro approach affects C++ symbols as well (albeit, pre-mangling, but if A and B are distinct symbols pre-mangling, their post-mangling representations are also distinct, so it still accomplishes our goal).

C++ and macros don't mix well. The way to do C++ is an inline namespace. But this is easy and very straightforward. (If you only use libcrypto, you don't care about C++ for now. But we may add a C++ dependency later.)

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>