Sync bundled bits of golang.org/x/crypto.
We no longer need to fork them. This is in preparation for pulling it
via Go modules, but probably need to figure out the network issue first.
Slightly bad manners for CI to do that.

As before, I've patched out the Poly1305 assembly, just to avoid
carrying stuff we don't need.

Change-Id: Ic258264f3c3559817d5e4921e4ad3282e94d05fe