Add an osx_sdk cache to BoringSSL builders.
This will carry the hermitic toolchain for Mac builders. It was
suggested that using mixins rather than adding empty caches to each
builder was preferable, so this also makes the try builders use mixins.

This CL adds a CPU selector to luci.boringssl.ci to match
luci.boringssl.try. It's a no-op, but we may as well note that we do
indeed want an x86_64 machine.

It also aligns the macOS versions a bit. The try jobs were using a mix
of 10.9 and 10.12 because, at the time, iOS required 10.12 but 10.9 was
more plentiful. Now 10.13 is more plentiful. Using the mixin is probably
better so we can update it all at once. (There's something to be said
for running the tests on older versions, but we haven't been doing that
historically. That kind of coverage is probably more efficient to do on
the embedder side like Chromium.)

There's a similar version disparity on Windows between Win7 and Win10.
I've left those alone in this CL and added a "win7" mixin.

Bug: chromium:890351
Change-Id: I11bbddaa513ec5b48dacb2626bd7b282473d4a43