Remove cpu/cores filter from Android bots.
This is a speculative fix. I thought it was the cache filter, but
there's a default cpu and cores filter being applied and the Android
bots don't have core and cores set at all (is that a mistake?).

An earlier change moved the cpu filter into the mixins to align the CI
and CQ, so we can just remove that. Remove the cores filter at all, to
further align them. The luci.flex.ci pool doesn't have variations in the
number of cores (aside from Mac which already cancelled the filter), and
we don't particularly care how many cores.

Bug: chromium:890955
Change-Id: I9da9b6513da93ea861bb0ab4aa1a6daa8fa35eba