[perlasm] Don't erase externs in ARM assembly
This CL removes the last line of the extern subroutine in arm-xlate.pl
that disabled output for .extern directives.  With that line in place
the generated assembly would redefine symbols like OPENSSL_armcap_P and
lose their hidden visibility.  This becomes a problem when standalone
lld tries to link a static BoringSSL (as part of a --start/end-group)
into a shared object.

Change-Id: I524b3dce7e5cfe017498847fbf9b8a5df4b98fce