Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > (1 comment)
> 
> Well, at least that gives some clarity as to why it was the way it was.  I'll try fiddling ome more with it in a bit.  I apologize that I don't have a good iOS build env to try these things prior to the tryjobs.

No worries! That's what try jobs are for. We usually do the same. :-)

(If you have a Mac, there's some notes on how to compile it for iOS in BUILDING.md. As for how to run it... we, uh, haven't figured that out yet either! Actually running it has to wait until it rolls into Chrome/iOS's waterfall right now, which is not ideal but what we've got...)

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>