Update patch set 4
Patch Set 4: Code-Review+2

Ohhh. Okay, I think the problem is which things do and don't count the hash-inserted padding are a little fuzzy. Also, depending on what we're measuring, we're actually subtracting a floor and a ceiling, which gives a +1.

Meh. Seems reasonable. It's a bit late for me to stare at things intently, and we only actually use the HMAC-SHA1 ones, for which this gives the same answer. :-)

One of these days, I'll get around to implementing this for Go and maybe I'll manage to find a less confusing formulation...

Patch-set: 4
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=+2