Make SSL_get_current_cipher valid during QUIC callbacks.
Update-Note: This effectively reverts https://boringssl-review.googlesource.com/4733, which was an attempt at a well-defined story during renegotiation and pre-handshake. This is a behavior change, though one that matches OpenSSL upstream. It is also more consistent with other functions, such as SSL_get_curve_id. Renegotiation is now opt-in, so this is less critical, and, if we change the behavior mid-renegotiation, we should do it consistently to all getters.

Change-Id: Ica6b386fb7c5ac524395de6650642edd27cac36f