Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> FWIW, I have some C code sitting around that does the inversion faster (IIRC) than that assembly code does it. If you split that change out then maybe it would save some effort for me to submit that C code to use instead. The main thing that's blocking it is the lack of review. (My C code uses a more efficient algorithm that should be much more efficient if/when it is translated into assembly without memory accesses.)

I'm going to try and get this relanded as-is in order to move forward. However, if you have something better please do feel free to upload it in any form that's easy for you. I.e. just add a random .c file, paste it in, and upload it as a CL. We might take a while to get around to it, but we'll probably integrate it eventually!

Patch-set: 2