Update patch set 2
Patch Set 2: Code-Review-2

BoringSSL should not be packaged individually because it doesn't have a stable ABI. Nor is it intended for use outside of Google. To quote our site: "Although BoringSSL is an open source project, it is not intended for general use, as OpenSSL is. We don't recommend that third parties depend upon it."

Patch-set: 2
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=-2