Update patch set 2
Patch Set 2: Code-Review-1

> We intentionally did the negotiation independent of key_shares because there was a lot of unclear bits around how the client filled in key_shares. There's some machinery in the spec (e.g. server-sent supported_groups) which allows for clients to be stateful here, which can have some downgrade-smelling consequences for a server that does negotiation based on key_shares.
> 
> For instance, consider a client which remembered the previously used group and always predicted that one. A server with this CL's behavior would then get stuck at an older group. One can fix this with the server-sent supported_groups extension that the client would then evaluate, but we don't send that (though we could) and it is somewhat counter to TLS's usual client-advertise/server-choose pattern. The chooser half is free to implement more complex selection algorithms (e.g. equipreference groups) that don't fit in a server-sent list.
> 
> Granted, we're basically the only implementation that took this view. Though I do still think this view makes more sense. :-) What scenario were you thinking wanted a key_share-based algorithm?

I was just in the area because of CECPQ2 (which will be coming your way for an initial review today—it's huge, but it's about ready for initial review.)

I had not considered that clients might do something dynamic but, now that you mention it, I can see that sort of thing happening and having them be stuck on a sub-optimal curve would be bad. In a world where clients are static, I think this change makes sense, but you're right about the risks.

We are aware of clients that have state like that? Do they follow a server's supported_groups if given?

Patch-set: 2
Reviewer: Gerrit User 5070 <5070@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=-1
Label: Commit-Queue=0