Update patch set 1
Patch Set 1:

BTW, my understanding is that only the VPAES and `aes_hw` implementations of `AES_encrypt` are side-channel-free with respect to memory access patterns, and VPAES and BSAES cannot be mixed together (on some platforms? on all platforms?). This makes it more important to use AES-NI when it is available.

Patch-set: 1
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>