Update patch set 2
Patch Set 2:

(3 comments)

Removing BSAES is going to take some measurement and thought because AESNI-less machines are, sadly, reasonably common, and BSAES was a perf win last I measured. :-( Though I too am not happy about BSAES's internal fallback to the table-based AES, so it's worth looking at that again.

But in the meantime, this change is possible on its own without much thought, and useful for QUIC packet number encryption.

(If you're not interested in doing the rebase, I'm happy to put it together myself. Comments below are partially with the rebase in mind.)

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>