Update patch set 2
Patch Set 2:

Note that this may not be exactly the right thing to do:

Check out this code in aesni-x86_64.pl

```
	leaqbit	OPENSSL_ia32cap_P(%rip),%r10
	movl	4(%r10),%r10d
	and	\$`1<<28|1<<11`,%r10d	# AVX and XOP bits
	
.L10rounds:
	mov	\$9,$bits			# 10 rounds for 128-bit key
	cmp	\$`1<<28`,%r10d			# AVX, bit no XOP
	je	.L10rounds_alt
```

Perhaps this code is testing the XOP bit bere because the AVX code path isn't faster than its AESNI code on AMD processors. However, if that's the case then IMO it shouldn't be testing the XOP bit, but instead the Intel bit.

Patch-set: 2
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>