Update patch set 3
Patch Set 3:

We did benchmarks of VPAES vs BSAES on the sort of older hardware that would use it. BSAES is 50% faster for sizes that the network stack tends to use, so it's not trivial. However, it isn't ok that the fallback isn't constant time so we're going to look at fiddling the key-schedule on the fly when we need to fallback in order to use VPAES.

Patch-set: 3
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>