Update patch set 2
Patch Set 2:

At a higher level:

The client side of this change would need to be informed by how we envision implementing with external verifiers. Do you need the client-side of it at this time?

Also David would need to weigh in because I think he's been thinking about integrated multi-cert support and this might impact that.

Patch-set: 2